+++ visualization/app/codeCharta/ui/toolBar/presentationModeButton/presentationModeButton.component.spec.ts
@@ -1,0 +1,21 @@
+import { TestBed } from "@angular/core/testing"
+import { render, screen } from "@testing-library/angular"
+import userEvent from "@testing-library/user-event"
+import { PresentationModeButtonComponent } from "./presentationModeButton.component"
+import { PresentationModeButtonModule } from "./presentationModeButton.module"
+
+describe("presentationModeButtonComponent", () => {
+	beforeEach(() => {
+		TestBed.configureTestingModule({
+			imports: [PresentationModeButtonModule]
+		})
+	})
+
+	it("should toggle presentation mode", async () => {
+		const { container } = await render(PresentationModeButtonComponent, { excludeComponentDeclaration: true })
+		expect(screen.getByTitle("Enable presentation mode")).toBeTruthy()
+
+		userEvent.click(container.querySelector("mat-slide-toggle label"))
+		expect(screen.getByTitle("Disable presentation mode")).toBeTruthy()
+	})
+})